# openapi.yaml
openapi: 3.1.0
info:
  title: Shoutboard API
  version: 1.0.0
  description: |
    API-first employee recognition group card platform. Shoutboard enables teams to
    create collaborative recognition cards for birthdays, promotions, farewells, and
    other occasions with public shareable links and team contributions.

    ## Key Features
    - **Multi-occasion support**: Birthday, anniversary, promotion, welcome, farewell, custom
    - **Public sharing**: Generate shareable links for public contribution
    - **Team collaboration**: Collect messages from multiple contributors anonymously or identified
    - **Webhook integrations**: Real-time event notifications for board creation, posting, and delivery
    - **API-first design**: Full programmatic access with JWT or API key authentication
    - **Analytics**: Track board engagement, contributor metrics, and team participation

    ## Authentication Methods
    The API supports two authentication methods:

    ### JWT Bearer Token
    Register and login to receive a JWT token valid for 7 days. Include in Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ### API Key
    Create persistent API keys in the dashboard for service-to-service integrations.
    Include in header:
    ```
    X-API-Key: sb_live_<random_string>
    ```
    API keys can be scoped to specific permissions and revoked at any time.

    ## Webhook Events
    Subscribe to real-time events:
...

...

    ## Rate Limiting
...

...
    ```json
...

...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: string
          description: Unique organization ID (CUID)
        name:
          type: string
          description: Organization display name
        slug:
          type: string
          description: URL-safe organization identifier
        plan:
          type: string
          enum: [STARTER, GROWTH, SCALE, ENTERPRISE]
          description: Subscription plan tier
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, slug, plan, createdAt, updatedAt]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [ADMIN, MEMBER]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, email, name, role, createdAt, updatedAt]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [success, error]
...
paths:
...
  /auth/register:
    post:
      summary: Register new organization and admin user
      operationId: registerUser
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orgName:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: Acme Corporation
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: Jane Smith
                email:
                  type: string
                  format: email
                  example: jane@acme.com
                password:
                  type: string
                  minLength: 8
                  format: password
                  example: SecurePassword123!
              required: [orgName, name, email, password]
      responses:
        201:
          description: Organization and user created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: JWT token valid for 7 days
                      user:
                        $ref: '#/components/schemas/User'
                      org:
                        $ref: '#/components/schemas/Organization'
                    required: [token, user, org]
                required: [success, data]
        422:
          description: Validation error or email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: VALIDATION_ERROR
                  message: Email already registered
  /auth/login:
    post:
      summary: Login with email and password
      operationId: loginUser
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: jane@acme.com
                password:
                  type: string
                  format: password
                  example: SecurePassword123!
              required: [email, password]
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      user:
                        $ref: '#/components/schemas/User'
                      org:
                        $ref: '#/components/schemas/Organization'
                required: [success, data]
        401:
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: INVALID_CREDENTIALS
                  message: Invalid email or password
  /auth/me:
    get:
      summary: Get current authenticated user
      operationId: getCurrentUser
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Current user information
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      org:
                        $ref: '#/components/schemas/Organization'
                required: [success, data]
        401:
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Board Endpoints
...